+++ memberaudit/tests/models/test_character_1.py
@@ -9,6 +9,7 @@ from esi.errors import TokenError
 from esi.models import Token
 from eveuniverse.models import EveEntity, EveMarketPrice, EveSolarSystem, EveType
 
+from allianceauth.eveonline.models import EveCharacter
 from app_utils.testing import NoSocketsTestCase, create_user_from_evecharacter
 
 from ...models import (
@@ -46,13 +47,13 @@ class TestCharacter(NoSocketsTestCase):
         super().setUpClass()
         load_entities()
 
-    def test_is_main_1(self):
+    def test_should_identify_main(self):
         # given
         character_1001 = create_memberaudit_character(1001)
         # when/then
         self.assertTrue(character_1001.is_main)
 
-    def test_is_main_2(self):
+    def test_should_be_true_for_main_only(self):
         # given
         character_1001 = create_memberaudit_character(1001)
         user = character_1001.eve_character.character_ownership.user
@@ -61,7 +62,7 @@ class TestCharacter(NoSocketsTestCase):
         self.assertTrue(character_1001.is_main)
         self.assertFalse(character_1101.is_main)
 
-    def test_is_main_3(self):
+    def test_should_be_false_when_no_main(self):
         # given
         character_1001 = create_memberaudit_character(1001)
         user = character_1001.eve_character.character_ownership.user
@@ -69,6 +70,12 @@ class TestCharacter(NoSocketsTestCase):
         user.profile.save()
         # when/then
         self.assertFalse(character_1001.is_main)
+
+    def test_should_be_false_when_orphan(self):
+        # given
+        character = create_character(EveCharacter.objects.get(character_id=1121))
+        # when/then
+        self.assertFalse(character.is_main)
 
     def test_should_keep_sharing(self):
         # given
@@ -84,6 +91,27 @@ class TestCharacter(NoSocketsTestCase):
         # then
         character.refresh_from_db()
         self.assertTrue(character.is_shared)
+
+    def test_should_identify_user_of_a_character(self):
+        # given
+        character_1001 = create_memberaudit_character(1001)
+        user = character_1001.eve_character.character_ownership.user
+        # when/then
+        self.assertTrue(character_1001.user_is_owner(user))
+
+    def test_should_identify_not_user_of_a_character(self):
+        # given
+        character_1001 = create_memberaudit_character(1001)
+        user = create_user_from_evecharacter(1002)
+        # when/then
+        self.assertFalse(character_1001.user_is_owner(user))
+
+    def test_should_identify_not_user_of_an_orphan(self):
+        # given
+        character = create_character(EveCharacter.objects.get(character_id=1121))
+        user = create_user_from_evecharacter(1002)
+        # when/then
+        self.assertFalse(character.user_is_owner(user))
 
     def test_should_remove_sharing(self):
         # given
@@ -258,31 +286,44 @@ class TestCharacterFetchToken(TestCase):
         super().setUpClass()
         load_entities()
 
-    def setUp(self) -> None:
-        self.character = create_memberaudit_character(1001)
-
-    def test_defaults(self):
-        token = self.character.fetch_token()
+    def test_should_return_token_with_default_scopes(self):
+        # given
+        character = create_memberaudit_character(1001)
+        # when
+        token = character.fetch_token()
+        # then
         self.assertIsInstance(token, Token)
         self.assertSetEqual(scope_names_set(token), set(Character.get_esi_scopes()))
 
-    def test_specified_scope(self):
-        token = self.character.fetch_token("esi-mail.read_mail.v1")
+    def test_should_return_token_with_specified_scope(self):
+        # given
+        character = create_memberaudit_character(1001)
+        # when
+        token = character.fetch_token("esi-mail.read_mail.v1")
         self.assertIsInstance(token, Token)
         self.assertIn("esi-mail.read_mail.v1", scope_names_set(token))
 
-    @patch(MODELS_PATH + ".character.notify_throttled")
-    def test_should_raise_exception_and_notify_user_if_not_found(
-        self, mock_notify_throttled
-    ):
+    def test_should_raise_exception_with_scope_not_found_for_orphans(self):
+        # given
+        character = create_character(EveCharacter.objects.get(character_id=1121))
         # when
         with self.assertRaises(TokenError):
-            self.character.fetch_token("invalid_scope")
+            character.fetch_token()
+
+    @patch(MODELS_PATH + ".character.notify_throttled")
+    def test_should_raise_exception_and_notify_user_if_scope_not_found(
+        self, mock_notify_throttled
+    ):
+        # given
+        character = create_memberaudit_character(1001)
+        # when
+        with self.assertRaises(TokenError):
+            character.fetch_token("invalid_scope")
         # then
         self.assertTrue(mock_notify_throttled.called)
         _, kwargs = mock_notify_throttled.call_args
         self.assertEqual(
-            kwargs["user"], self.character.eve_character.character_ownership.user
+            kwargs["user"], character.eve_character.character_ownership.user
         )