+++ memberaudit/views/character_viewer_1.py
@@ -140,35 +140,7 @@ def character_viewer(request, character_pk: int, character: Character) -> HttpRe
         }
     )
 
-    # list of all characters owned by this user for sidebar
-    accessible_characters = set(
-        Character.objects.user_has_access(user=request.user).values_list(
-            "pk", flat=True
-        )
-    )
-    all_characters = (
-        EveCharacter.objects.select_related(
-            "character_ownership__memberaudit_character"
-        )
-        .filter(
-            character_ownership__user=character.eve_character.character_ownership.user
-        )
-        .order_by("character_name")
-        .annotate(memberaudit_character_pk=F("memberaudit_character"))
-        .annotate(is_shared=F("memberaudit_character__is_shared"))
-        .values(
-            "character_id", "character_name", "memberaudit_character_pk", "is_shared"
-        )
-    )
-    all_characters = [
-        {
-            **obj,
-            **{
-                "has_access": (obj["memberaudit_character_pk"] in accessible_characters)
-            },
-        }
-        for obj in all_characters
-    ]
+    all_characters = _identify_user_characters(request, character)
 
     # assets total value
     character_assets_total = (
@@ -229,6 +201,43 @@ def character_viewer(request, character_pk: int, character: Character) -> HttpRe
         "memberaudit/character_viewer.html",
         add_common_context(request, context),
     )
+
+
+def _identify_user_characters(request, character):
+    """Identify all characters owned by this user for siderbar."""
+
+    try:
+        user = character.eve_character.character_ownership.user
+    except ObjectDoesNotExist:
+        eve_characters_of_user = EveCharacter.objects.none()
+    else:
+        eve_characters_of_user = EveCharacter.objects.select_related(
+            "character_ownership__memberaudit_character"
+        ).filter(character_ownership__user=user)
+
+    all_characters = (
+        eve_characters_of_user.order_by("character_name")
+        .annotate(memberaudit_character_pk=F("memberaudit_character"))
+        .annotate(is_shared=F("memberaudit_character__is_shared"))
+        .values(
+            "character_id", "character_name", "memberaudit_character_pk", "is_shared"
+        )
+    )
+    accessible_characters = set(
+        Character.objects.user_has_access(user=request.user).values_list(
+            "pk", flat=True
+        )
+    )
+    all_characters = [
+        {
+            **obj,
+            **{
+                "has_access": (obj["memberaudit_character_pk"] in accessible_characters)
+            },
+        }
+        for obj in all_characters
+    ]
+    return all_characters
 
 
 @login_required